lib/string/: Move string-related files to string/ subdir
