+++ lib/copydir.c
@@ -36,7 +36,7 @@
 #include <attr/libattr.h>
 #endif				/* WITH_ATTR */
 #include "shadowlog.h"
-#include "sprintf.h"
+#include "string/sprintf.h"
 
 
 static /*@null@*/const char *src_orig;
