+++ lib/string/stpecpy.h
