+++ lib/string/strncpy.h
