+++ lib/string/strtcpy.h
