+++ lib/string/zustr2stp.h
