+++ src/useradd.c
@@ -63,7 +63,7 @@
 #include "tcbfuncs.h"
 #endif
 #include "shadowlog.h"
-#include "sprintf.h"
+#include "string/sprintf.h"
 
 
 #ifndef SKEL_DIR
